requirements.txt
beautifulsoup4==4.13.4
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
nltk==3.8.1
wordcloud==1.9.2
flask==2.3.3
snowflake-connector-python==3.0.4 